Pull container view into separate class
